fix: 디렉터리 구조 수정, use form 사용(#19)
